// apps/web/src/shared/shared.enum.ts
export enum StatusEnum {
  NOT_STARTED = 'not_started',
  STUCK = 'stuck',
  WORKING_ON_IT = 'working_on_it',
  DONE = 'done'
}

export enum PriorityEnum {
  LOW = 'low',
  MEDIUM = 'medium',
  HIGH = 'high',
  CRITICAL = 'critical'
}

export enum TableColorEnum {
  RED = '#ff5733',
  ORANGE = '#ff8c00',
  YELLOW = '#ffd700',
  GREEN = '#32cd32',
  BLUE = '#1e90ff',
  PURPLE = '#9370db',
  PINK = '#ff69b4',
  CYAN = '#00ced1',
  LIME = '#00ff00',
  MAGENTA = '#ff00ff',
  BROWN = '#8b4513',
  GRAY = '#808080',
  BLACK = '#000000',
  WHITE = '#ffffff',
  DEFAULT = '#6b7280'
}